Add files via upload
It's been a great time learning with Harvoxx school. I aim to become a renowned Cybersecurity expert and I'm focused on taking a step at a time. Understanding how to manipulate files using kali linux commands and accomplishing tasks is one of my wins as a beginner and I'm glad to share my progress.